Restructuring the pathfinding classes
