+++ Bardez.Project.SwordOfTheStars.IO/Pathfinding/Pathfinder.cs
@@ -1,12 +1,13 @@
 ﻿using System;
-using Bardez.Project.SwordOfTheStars.IO.Steam;
+using Bardez.Project.SwordOfTheStars.IO.Pathfinding.Physical;
+using Bardez.Project.SwordOfTheStars.IO.Pathfinding.Steam;
 
-namespace Bardez.Project.SwordOfTheStars.IO;
+namespace Bardez.Project.SwordOfTheStars.IO.Pathfinding;
 
 /// <summary>Class that should return the path to the root Sword of the Stars directory</summary>
 public class Pathfinder
 {
-    public String DeriveSotsPath()
+    public string DeriveSotsPath()
     {
         //first try to fetch from the registry. Old-school disks install.
         var sotsPath = RegistryPathfinder.ReadSotsPath();
